+++ lib/langchain/active_record/hooks.rb
@@ -35,7 +35,7 @@ module Langchain
     # Query the vector search provider
     #     Recipe.similarity_search("carnivore dish")
     # Delete the default schema to start over
-    #     Recipe.class_variable_get(:@@provider).client.schema.delete class_name: "Recipes"
+    #     Recipe.class_variable_get(:@@provider).destroy_default_schema
     #
     module Hooks
       def self.included(base)
@@ -87,7 +87,9 @@ module Langchain
             query: query,
             k: k
           )
-          ids = records.map { |record| record.dig("__id") }
+
+          # We use "__id" when Weaviate is the provider
+          ids = records.map { |record| record.dig("id") || record.dig("__id") }
           where(id: ids)
         end
       end